initial commit of fixed wind function. added a custom function changed simulation of routes to continue reducing SD as got closer to end
Removed big chunks of code which I originally wrote to do the geolocation error script. Instead, I ran the simulation of geolocation error inside the loop of the real birds. This is because the error is to do with the real birds' tracks, not the tracks of the simulated individuals.

Minor edits to file paths